// src/app/pages/AboutMe/types.d.ts
export interface IPropsContainer {}

export interface IPropsPresentation {
  colorTitle: string;
  linkToLinkedin: string;
  colorOthersText: string;
  colorParagraph: string;
  aboutMe: string[];
}

export interface ITitlePage {
  color: string;
}

export interface IContainerIconProps {
  color: string;
}

export interface IContainerOthers {
  color: string;
}

export interface IContentText {
  color: string;
}